% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{PermSumCount.multipleX.test}
\alias{PermSumCount.multipleX.test}
\title{Function for performing the Wilcoxon rank test, over multiple rarefactions of the data.
Pvalue is computed by averaging over the P.values of multiple rarefactions.
WARNING: This function was used to test the this approach is invalid - do not use in practice.
See paper for further details}
\usage{
PermSumCount.multipleX.test(List_of_X, Y, B = as.integer(c(1000)),
  is.Asymptotic = F, disable.ties.correction = F)
}
\arguments{
\item{List_of_X}{list of vectors, corresponding to multiple rarefactions}

\item{Y}{vector of 0/1's - group labeling}

\item{B}{number of permutations}

\item{is.Asymptotic}{logical - should use asymptotic P-values}

\item{disable.ties.correction}{logical - should the correction for ties be disabled}
}
\value{
list with a single item - the average P.value across different rarefactions
}
\description{
Function for performing the Wilcoxon rank test, over multiple rarefactions of the data.
Pvalue is computed by averaging over the P.values of multiple rarefactions.
WARNING: This function was used to test the this approach is invalid - do not use in practice.
See paper for further details
}
