% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{PermSumCount.test}
\alias{PermSumCount.test}
\title{Permutation based wilcoxon rank sum test
Function for performing the Wilcoxon rank sum test, with permutations}
\usage{
PermSumCount.test(X, Y, B = as.integer(c(40000)),
  DoWald = as.integer(c(1)), return_perms = F,
  disable.ties.correction = F)
}
\arguments{
\item{X}{a vector of counts}

\item{Y}{a vector of 0,1's - group labeling}

\item{B}{number of permutations}

\item{DoWald}{Should an early stopping rule be used}

\item{return_perms}{should the permutation P-values be given}

\item{disable.ties.correction}{should the correction for ties be disabled}
}
\value{
a list with the test statistic, P.value (2-sided), number of permutations and the asymptotic P.value. Item perms will be included if permutations were requested
}
\description{
Permutation based wilcoxon rank sum test
Function for performing the Wilcoxon rank sum test, with permutations
}
